requirements.txt
selenium==4.15.0
pytest==7.4.3
allure-pytest==2.13.2
pytest-rerunfailures==12.0
webdriver-manager==4.0.1